Moving all entities to entities dir
