+++ src/js/app.js
@@ -1,7 +1,7 @@
 import {Game} from './game.js'
 import {Drawer} from './drawer.js'
 import {Ticker} from './ticker.js'
-import {Player} from './player.js'
+import {Player} from './entities/player.js'
 import {World} from './world.js'
 import {TileMap} from './tile_map.js'
 import {TileMapDrawer} from './drawers/tile_map_drawer.js'
