+++ src/js/entities/entity.js
