+++ src/js/entities/player.js
